# Eliminate write barrier branches from codegen and late-gc-lowering