--- a/src/aotcompile.cpp
+++ b/src/aotcompile.cpp
@@ -694,7 +694,7 @@ void addOptimizationPasses(legacy::PassManagerBase *PM, int opt_level,
     // Running `memcpyopt` between this and `sroa` seems to give `sroa` a hard time
     // merging the `alloca` for the unboxed data and the `alloca` created by the `alloc_opt`
     // pass.
-    PM->add(createArraylenOptPass());
+    PM->add(createArrayOptPass());
     PM->add(createAllocOptPass());
     // consider AggressiveInstCombinePass at optlevel > 2
     PM->add(createInstructionCombiningPass());
@@ -741,6 +741,7 @@ void addOptimizationPasses(legacy::PassManagerBase *PM, int opt_level,
     PM->add(createSimpleLoopUnrollPass());
 
     // Run our own SROA on heap objects before LLVM's
+    PM->add(createArrayOptPass());
     PM->add(createAllocOptPass());
     // Re-run SROA after loop-unrolling (useful for small loops that operate,
     // over the structure of an aggregate)
@@ -771,7 +772,6 @@ void addOptimizationPasses(legacy::PassManagerBase *PM, int opt_level,
 
     // More dead allocation (store) deletion before loop optimization
     // consider removing this:
-    PM->add(createArraylenOptPass());
     PM->add(createAllocOptPass());
     // see if all of the constant folding has exposed more loops
     // to simplification and deletion