--- a/src/cgutils.cpp
+++ b/src/cgutils.cpp
@@ -2475,8 +2475,8 @@ static Value *emit_arraylen_prim(jl_codectx_t &ctx, const jl_cgval_t &tinfo)
                 jl_alloc::AllocIdInfo info;
                 if (jl_alloc::getArrayAllocInfo(info, alloc)) {
                     assert(static_cast<size_t>(info.array.dimcount) == ndim);
-                    llvm::Value *length = ConstantInt::get(getSizeTy(ctx.builder.getContext()), 1);
-                    for (unsigned i = 0; i < ndim; i++) {
+                    llvm::Value *length = alloc->getArgOperand(ndim);
+                    for (unsigned i = ndim - 1; i --> 0;) {
                         length = ctx.builder.CreateMul(length, alloc->getArgOperand(i + 1), "", true, true);
                     }
                     return length;