--- a/src/llvm-alloc-helpers.cpp
+++ b/src/llvm-alloc-helpers.cpp
@@ -303,20 +303,28 @@ void jl_alloc::runEscapeAnalysis(llvm::Instruction *I, EscapeAnalysisRequiredArg
     }
 }
 
-//FIXME: This doesn't actually work on Windows, as Windows inserts a trampoline
-//function which doesn't have the same function pointer as the bare alloc_array
-//functions
+//We use metadata installed during codegen to identify array allocations
+//This is because codegen creates naked function pointer calls and thus
+//strips the identifying information about what the underlying function
+//actually is, so we need codegen to pass along the lost information.
 bool jl_alloc::getArrayAllocInfo(AllocIdInfo &info, llvm::CallInst *call) {
-    if (call->getMetadata("allocation.array")) {
-        info.isarray = true;
-        info.array.dimcount = call->getNumArgOperands() - 1;
-        return true;
-    } else {
-        if (call->getMetadata("allocation.array.dyn")) {
-            info.isarray = true;
-            info.array.dimcount = 0;
-            return true;
+    if (auto md = call->getMetadata("julia.array")) {
+        if (md->getNumOperands() == 1) {
+            if (auto mds = dyn_cast<MDString>(md->getOperand(0).get())) {
+                if (mds->getString() == "allocation") {
+                    info.isarray = true;
+                    info.array.dimcount = call->arg_size() - 1;
+                    return true;
+                } else if (mds->getString() == "allocation.dyn") {
+                    info.isarray = true;
+                    info.array.dimcount = 0;
+                    return true;
+                } else {
+                    assert(false && "Expected julia.array metadata to be either 'allocation' or 'allocation.dyn'!");
+                }
+            }
         }
+        assert(false && "Expected julia.array metadata to have one string operand!");
     }
     return false;
 }