… src/llvm-alloc-helpers.h
… src/llvm-array-opt.cpp
… src/llvm-arraylen-opt.cpp
… src/llvm-late-gc-lowering.cpp
… test/llvmpasses/array-opt.ll