Branches on the how flag of freshly allocated arrays for write barriers are unnecessary and could block escape analysis of arrays in the future. This PR proposes to eliminate those branches as generated in codegen and late-gc-lowering.

Compilation timing comparison:
```
Sysimage built. Summary:
Total ───────  61.639983 seconds
Base: ───────  24.926984 seconds 40.4396%
Stdlibs: ────  36.711467 seconds 59.5579%
    JULIA usr/lib/julia/sys-o.a
Generating REPL precompile statements... 36/36
Executing precompile statements... 1393/1428
Precompilation complete. Summary:
Total ─────── 115.848746 seconds
Generation ──  89.879481 seconds 77.5835%
Execution ───  25.969266 seconds 22.4165%
```
PR:
```
Sysimage built. Summary:
Total ───────  60.190534 seconds
Base: ───────  24.592287 seconds 40.8574%
Stdlibs: ────  35.596712 seconds 59.1401%
    JULIA usr/lib/julia/sys-o.a
Generating REPL precompile statements... 36/36
Executing precompile statements... 1393/1428
Precompilation complete. Summary:
Total ─────── 114.047976 seconds
Generation ──  87.002346 seconds 76.2857%
Execution ───  27.045631 seconds 23.7143%
```

Godbolt comparison:
Master: https://godbolt.org/z/5KrTqnW4e
PR: https://godbolt.org/z/oP6z4nT5a

This PR depends on #43547 and #43487 for array optimization improvements and allocation identification, and #43057 for escape analysis improvements.