@pchintalapudi 2022-01-05
Branches in late-gc-lower cannot be removed, as they are dependent on unobservable GC low bit stores and cannot be forwarded from the original type pointer.